Added password validator on settings and child creation Let user unassign cards/categories assigned by others

+++ helper/formatDate.ts
@@ -5,7 +5,7 @@ export const formatDate = (date: Date | string | null): string => {
   const d = new Date(date);
 
   const day = String(d.getDate()).padStart(2, "0"); // 01 → 09
-  const month = d.toLocaleString("default", { month: "short" });
+  const month = d.toLocaleString("default", { month: "long" });
   const year = d.getFullYear();
 
   return `${month} ${day}, ${year}`;
